chore: move schema generation to a single file
