+++ packages/core/src/server/generateGraphqlSchemaFile.ts
@@ -1,10 +1,73 @@
-import { GraphQLSchema } from 'graphql'
 import { writeFileSync } from 'fs-extra'
+import { getSchema, getTypeDefs } from '@faststore/api'
 import { printSchemaWithDirectives } from '@graphql-tools/utils'
+import { loadFilesSync } from '@graphql-tools/load-files'
+import { mergeTypeDefs } from '@graphql-tools/merge'
+import { buildASTSchema } from 'graphql'
+import type { GraphQLSchema } from 'graphql'
+import type { TypeSource } from '@graphql-tools/utils'
 
-import { apiSchema } from './schema'
+import { apiOptions } from './options'
 
-async function createGraphqlSchemaFile(apiSchema: Promise<GraphQLSchema>) {
+export function getTypeDefsFromFolder(
+  customPath: string | string[]
+): TypeSource {
+  const basePath = ['src', 'customizations', 'graphql']
+
+  const pathArray = Array.isArray(customPath) ? customPath : [customPath]
+
+  return (
+    loadFilesSync([...basePath, ...pathArray, 'typeDefs'], {
+      extensions: ['graphql'],
+    }) ??
+    loadFilesSync([...basePath, ...pathArray, 'typedefs'], {
+      extensions: ['graphql'],
+    })
+  )
+}
+
+export function getCustomSchema(typeDefs: TypeSource[]) {
+  try {
+    const mergedTypeDefs = mergeTypeDefs(typeDefs)
+
+    const schema = buildASTSchema(mergedTypeDefs)
+
+    return schema
+  } catch (error) {
+    console.error(
+      `
+      An error occurred while attempting to merge the GraphQL Schema Extensions. Check the custom typeDefs and resolvers located in the 'customizations/graphql/' directory. The changes since the last successful schema merge will be ignored.
+
+      Error message:`,
+      error
+    )
+  }
+}
+
+export function getNativeTypeDefs() {
+  return getTypeDefs() as TypeSource
+}
+
+export function getVTEXExtensionsTypeDefs() {
+  return getTypeDefsFromFolder('vtex')
+}
+
+export function getThirdPartyExtensionsTypeDefs() {
+  return getTypeDefsFromFolder('thirdParty')
+}
+
+export const nativeApiSchema = getSchema(apiOptions)
+
+export const getMergedSchema = async () =>
+  getCustomSchema(
+    [
+      getNativeTypeDefs(),
+      getVTEXExtensionsTypeDefs(),
+      getThirdPartyExtensionsTypeDefs(),
+    ].filter(Boolean)
+  )
+
+async function writeGraphqlSchemaFile(apiSchema: Promise<GraphQLSchema>) {
   try {
     writeFileSync(
       [process.cwd(), '@generated', 'graphql', 'schema.graphql'].join('/'),
@@ -19,4 +82,7 @@ async function createGraphqlSchemaFile(apiSchema: Promise<GraphQLSchema>) {
   }
 }
 
-createGraphqlSchemaFile(apiSchema)
+// Schema with no resolvers - used to generate schema.graphql file
+const mergedApiSchema = getMergedSchema()
+
+writeGraphqlSchemaFile(mergedApiSchema)
